+++ delivery/src/main/java/com/eleven/logistics/delivery/presentation/dtos/CreateDeliveryRequest.java
@@ -1,4 +1,4 @@
-package com.eleven.logistics.delivery.presentation.dtos.delivery;
+package com.eleven.logistics.delivery.presentation.dtos;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
@@ -31,14 +31,14 @@ public class CreateDeliveryRequest {
   private UUID receiverSnsId;
 
   @NotNull
-  private UUID companyDeliveryManagerId;
+  private UUID companyDeliveryPersonId;
 
   private List<DeliveryRouteRequest> routes;
 
   public DeliveryRequest toDeliveryRequestDto() {
     return new DeliveryRequest(orderId, departureHubId, destinationHubId,
         deliveryAddress, receiver, receiverSnsId,
-        companyDeliveryManagerId);
+        companyDeliveryPersonId);
   }
 
   public List<DeliveryRouteRequest> getRouteDtos() {
