+++ delivery/src/main/java/com/eleven/logistics/delivery/presentation/dtos/DeliveryRequest.java
@@ -1,4 +1,4 @@
-package com.eleven.logistics.delivery.presentation.dtos.delivery;
+package com.eleven.logistics.delivery.presentation.dtos;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
@@ -30,5 +30,5 @@ public class DeliveryRequest {
   private UUID receiverSnsId;
 
   @NotNull
-  private UUID companyDeliveryManagerId;
+  private UUID companyDeliveryPersonId;
 }
