+++ delivery/src/main/java/com/eleven/logistics/delivery/presentation/dtos/DeliveryRouteRequest.java
@@ -1,4 +1,4 @@
-package com.eleven.logistics.delivery.presentation.dtos.delivery;
+package com.eleven.logistics.delivery.presentation.dtos;
 
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
